Chore: Made refactoring changes
